requirements.txt
python-telegram-bot==22.7
aiosqlite==0.19.0
matplotlib
python-dotenv==1.0.0
